+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## [1.0.0] - 2024-12-07
 ### Added
 - Config base files for the module
+
+## [1.2.0] - 2024-12-10
+### Added
+- New cron job for sync inventory with external system
+- Pre-configuration di.xml file for the module
+
+## [1.3.0] - 2024-12-11
+### Added
+-- New graphql query for getting product inventory with sku's